geoguard/requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
pyyaml
spacy
pytest
numpy
streamlit
plotly
